+++ src/app/auth/pages/signin/signin.component.html
@@ -6,14 +6,19 @@
     <!--inputs-->
     <mat-form-field appearance="outline" class="w-full">
       <mat-label>Email</mat-label>
-      <input matInput type="text" placeholder="john.doe@gmail.com" class="w-full" />
+      <input
+        matInput
+        type="text"
+        placeholder="john.doe@gmail.com"
+        class="w-full" />
     </mat-form-field>
     <mat-form-field appearance="outline" class="w-full">
       <mat-label>Password</mat-label>
       <input matInput type="password" class="w-full" />
     </mat-form-field>
     <div class="my-2 w-full">
-      <button type="submit"
+      <button
+        type="submit"
         class="w-full py-2 px-3 rounded-xl transition-colors text-white bg-slate-500 hover:bg-slate-700 font-semibold">
         Sign In
       </button>
