+++ src/app/auth/pages/signin/signin.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { SignInComponent } from "./sign-in.component";
+import { SignInComponent } from "./signin.component";
 
 describe("SignInComponent", () => {
   let component: SignInComponent;
